Fix compile issue
When trying to grab the application_path, it tries to grab "sys._MEIPASS", which only exists after using pyinstaller. This causes a compile failure when trying to run from python. To fix this, I have made it so application_path will be set to the one directory back from the current path, only if it is running from a script.