Fix the inability to kill Explorer
For some reason, the old code only worked on the Surface, but not on any other computers. This doesn't even look to see if Explorer is running, and instead just tries to kill it, and if it's already dead, then it just continues. It's honestly probably a better solution then before but