Fix CMD window opening
When certain commands ran when the application was built, a CMD window would popup. This commit fixes that, by changing "os.system" to "subprocess.call", and detaching it from a window. I also slightly modified the code that closes the splash screen to fix it, as it broke after making the adjustments to the code as mentioned before.