Commit technically fully functional version
The first version of Pomoimprove that is fully functional, both normally and through pyinstaller! This version will go to the start screen, ask a couple of questions, then go from a 10 minute break period with YouTube, to a 1 minute Spotify music picking period (if they picked to sign into a Spotify account), then to a 10 minute work period, then to a 5 minute break period, then it will repeat those two over and over. Although it is fully functional, there is still a couple of features and bugs that need to be ironed out, so that version will be 1.0.